+++ resources/lang/cs/edu.php
@@ -1,0 +1,26 @@
+<?php
+
+return [
+    'filter' => [
+      'media_type' => 'formát',
+      'target_group' => 'cílová skupina',
+      'suitable_for_home' => 'vhodné i pro rodiče',
+      'keywords' => 'klíčová slova',
+    ],
+
+    'media_type' => [
+      'methodology' => 'metodika',
+      'worksheet' => 'pracovní list',
+      'video' => 'video',
+      'collection' => 'kolekce',
+      'workshop' => 'workshop',
+      'virtual_exhibition' => 'virtuální prohlídka',
+    ],
+    
+    'age_group' => [
+      '3-6' => '3—6 let',
+      '7-10' => '7—10 let',
+      '11-15' => '11—15 let',
+      '16-19' => '16—19 let',
+    ],
+];